Update interactiveplots2.R
Some input widgets provide choices for the user to select from. Others allow the user to provide their own input. For example, input_text() provides a text field for users to type input into. Instead of assigning input_text() a choices argument, you assign it a value argument: a character string to display when the plot first loads.

By default, input widgets return their values as character strings and numbers. To have a widget return its value as a variable name, you need to add the extra argument map = as.name.

For example, the text widget in the first challenge will pass the character string "black" to the fill argument, which is useful for setting. If we add map = as.name to the arguments of input_text(), the widget would return ~black which is useful for mapping (or would be if black were a real variable in mtcars):
Instructions

    Change the radiobuttons widget in the first command on the right to a text widget that displays the initial value "black". The wanted resulting plot can be found here.
    In the second graph on the right, map fill to a select box that returns variable names. The select box should use the label "Choose fill variable:" and should offer the choices created by names(mtcars), as shown here. Use the map argument inside input_select() to map the selection instead of setting it.